requirements.txt
torch==1.10.0
torchvision==0.11.1
Pillow==6.2.1
statsmodels==0.12.1
numpy==1.19.5
Flask==2.0.2
